dt-headers fetch tests
